+++ app/src/main/AndroidManifest.xml
@@ -13,14 +13,10 @@
         android:theme="@style/Theme.JubinAndChildren"
         tools:targetApi="31">
         <activity
-            android:name=".Main"
-            android:exported="true" />
-        <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
